Saner module loading
This also fixes a recursive dependency bug.  Not because of prepending
module.loaded[...] = ... to each file (but I like this style as a
safeguard anyway), but because of doing require("entities") instead of
require("entities.whatever") individually for all the entities.  Lua's
kind of bad here, if you ask me.