Change client/server to link/terminal
The terminal is a local endpoint (socket).  It encapsulates all the
logic for receiving messages and tracking remote connection states.  The
(optional, maybe not really useful in practice) link is a shortcut for
sending to a particular remote.

Basically, the terminal is the useful data structure here, and it lets
you listen for incoming links or just send to a particular host or all
linked hosts.  The separate link object just gives you an easy way to
send to only one host, like for a client.

This also updates the game code to use the new API, but I didn't really
think that part through, so it may change in the near future.