Start huge revamp of net code to add reliability
This adds the skeleton of reliability in n3.  Packets now go out on
different channels (some of which will maintain order on receipt, in the
future), and are acknowledged when received (though they aren't re-sent
yet).  You'll be able to optionally send packets that don't need
ack'ing.

This lays the foundation for future work, but doesn't actually improve
the protocol very much as of yet.