Upgrade to Lua 5.3
It's been out for a while, might as well keep current.

The C code is pretty easy--unsigneds just become integers, no big whoop.
The lua code is a little trickier since every math operation has to
(gets to?) take the new integer type into account.  I think I got it all
ok, might've even gotten a little pedantic in there... but if this isn't
the place for pedantry, what is?